Desktop/abtech/requirements.txt
asgiref==3.8.1
dj-database-url==2.2.0
Django==5.0.6
djangorestframework==3.15.1
PyMySQL==1.1.1
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.12.0
